{"type":"TI", "runs":"0"}

+++ Hangman.java
@@ -28,7 +28,7 @@ public class Hangman extends ConsoleProgram {
     	println("Welcome to Hangman!");
     	
     	// get a guessing word from the dictionary
-    	int word_index = rgen.nextDouble(0, 9)
+    	int word_index = (int)rgen.nextDouble(0, 9)
     	hangmanlecxicon.getWord(index);
     	
     	// constant check user guessing results and chances left
